moving entities from Util to Entites
